Created a step_engine abstraction in C
