+++ FluidNC/include/Driver/delay_usecs.h
@@ -1,4 +1,8 @@
-#include <cstdint>
+#include <stdint.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif
 
 extern uint32_t ticks_per_us;
 
@@ -11,3 +15,7 @@ void delay_us(int32_t us);
 int32_t usToCpuTicks(int32_t us);
 int32_t usToEndTicks(int32_t us);
 int32_t getCpuTicks();
+
+#ifdef __cplusplus
+}
+#endif
