+++ FluidNC/include/Driver/gpio_dump.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <Print.h>
+
+void gpio_dump(Print& out);
