+++ platformio.ini
@@ -121,6 +121,7 @@ build_src_filter = ${common_esp32_base.build_src_filter} ${common_bt.build_src_f
 extends = common_esp32_s3
 lib_deps = ${common.lib_deps}
 build_src_filter = ${common_esp32_base.build_src_filter}
+build_flags = ${common_esp32_base.build_flags}
 
 [env:wifi_s3]
 extends = common_esp32_s3
